♻️ 코드 리팩토링 : FileMetaData 클래스 리팩토링 - 패키지를 `com.realworld.application.file.dto`에서 `com.realworld.common.type.file`로 이동 - 필드명 변경 :   - `targetDirectory` → `directory` - `create` 메서드 수정:   - 파라미터 순서 및 이름 변경   - 파일 이름 및 콘텐츠 타입 생성을 `FileFormat` enum을 사용하도록 변경
Ref : #18
